+++ source/renderer/app/themes/daedalus/cardano.js
@@ -824,6 +824,8 @@ export const CARDANO_THEME_OUTPUT = {
     '--theme-wallet-restore-dialog-new-label-background-color':
       'rgba(32, 34, 37, 0.1)',
     '--theme-wallet-restore-dialog-new-label-color': '#5e6066',
+    '--theme-wallet-restore-dialog-step-walletType-hardwareWalletDisclaimer-text-color':
+      '#ea4c5b',
   },
 };