+++ source/renderer/app/themes/daedalus/dark-blue.js
@@ -826,6 +826,8 @@ export const DARK_BLUE_THEME_OUTPUT = {
     '--theme-wallet-restore-dialog-new-label-background-color':
       'rgba(233, 244, 254, 0.1)',
     '--theme-wallet-restore-dialog-new-label-color': '#e9f4fe',
+    '--theme-wallet-restore-dialog-step-walletType-hardwareWalletDisclaimer-text-color':
+      '#ea4c5b',
   },
 };