+++ source/renderer/app/themes/daedalus/incentivized-testnet.js
@@ -814,6 +814,8 @@ export const INCENTIVIZED_TESTNET_THEME_OUTPUT = {
     '--theme-wallet-restore-dialog-new-label-background-color':
       'rgba(255, 255, 255, 0.1)',
     '--theme-wallet-restore-dialog-new-label-color': '#ffffff',
+    '--theme-wallet-restore-dialog-step-walletType-hardwareWalletDisclaimer-text-color':
+      '#ea4c5b',
   },
 };